+++ src/pipeline/pipe.js
@@ -1,0 +1,5 @@
+import Pipeline from './Pipeline'
+
+export default function pipe (...middleware) {
+  return new Pipeline(middleware)
+}