+++ test/header.js
@@ -1,8 +1,8 @@
-import { describe, it } from 'mocha'
+import { describe, it, before } from 'mocha'
 import chai from 'chai'
 import chaiAsPromised from 'chai-as-promised'
 
 chai.use(chaiAsPromised)
 chai.should()
 
-export { describe, it }
+export { describe, it, before }